requirements.txt
astroid==2.2.5
asyncio==3.4.3
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.13
django-bootstrap4==0.0.8
django-channels==0.7.0
django-countries==5.3.3
django-heroku==0.3.1
django-tagging==0.4.6
django-tagging-autocomplete-new==0.6.2
django-widget-tweaks==1.4.5
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
psycopg2==2.8.3
pylint==2.3.1
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rope==0.14.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
whitenoise==4.1.2
wrapt==1.11.2
